Swapped space character (#112)
* Fixed floating action button link

When clicking on the FloatingActionButton link it would send you to "https://flet.dev/docs/controls/filledbutton/floatingactionbutton" this of course does not exist. I have changed the link to "/docs/controls/floatingactionbutton" which should therefore send you to "https://flet.dev/docs/controls/floatingactionbutton"

* Swapped space

An IDEOGRAPHIC SPACE (U+3000) was used instead of a regular space (U+0020). This is an invalid character, therefore making the code unable to be run.