guide note: when async meets routing
This might already be known in the dev team or advanced user group
but I think it would be useful.

I know the new user's perspective can be valuable to improve the
documentation's readability. When I read the async app part I
realize it's not so intuitive to adapt to multi-view apps
since there's no `View.add_async` counterpart.