Updated TextThemeStyle usage in documentation
This commit updates the usage of TextThemeStyle in the Flet documentation. The `style` property was being used to set the TextThemeStyle, which was causing a UserWarning. This has been corrected to use the `theme_style` property instead, as recommended by the warning message. This change will help users follow best practices and avoid warnings.