Update file-picker-and-uploads.md
I made changes to the code of the section which explains the FilePicker.on_result event handler as the code does not follow in continuity with the previous two sections. This causes the an error when all three section of the code combined while running the program. The error stems from the line file_picker = ft.FilePicker(on_result=file_selection) as the file_picker = ft.FilePicker() is already defined in the first section